Update plugin-level reference comment
The comments about citing the dev docs seem a bit strange in the template. This is something that developers should cite in their repo and any article(s) as a template for their plugin. But a plugin-level citation then passes this on to the users; so the comment about keeping this plugin-level citation implies that all users of that plugin should also be citing the dev docs. This does not seem appropriate, as the users did not use the dev docs in their research.